refactored the reasoner-validator.util module to partition code into a versioning package and the trapi.util package
