+++ tests/test_validate.py
@@ -4,8 +4,7 @@ import pytest
 
 from jsonschema.exceptions import ValidationError
 
-from reasoner_validator.trapi import TRAPIValidator
-from reasoner_validator.util import openapi_to_jsonschema
+from reasoner_validator.trapi import TRAPIValidator, openapi_to_jsonschema
 
 TEST_VERSIONS = "1", "1.2", "1.2.0", "1.3", "1.3.0"
